+++ src/main/java/com/holidaymini/external/nager/dto/CountryResponse.java
@@ -1,4 +1,4 @@
-package com.holidaymini.external.dto;
+package com.holidaymini.external.nager.dto;
 
 public record CountryResponse(
         String countryCode,
